pyproject.toml
[tool.poetry]
name = "bill-yield-agent"
version = "0.1.0"
description = "Bill Yield Agent for ETHGlobal Bangkok"
authors = ["Otto G <5030059+ottodevs@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "operate" }]
include = [
    "operate/data/acn/*",
    "operate/data/contracts/*",
    "operate/data/contracts/service_staking_token/*",
    "operate/data/contracts/service_staking_token/build/*",
]

[tool.poetry.scripts]
operate = "operate.cli:main"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
open-autonomy = "==0.16.1"
open-aea-ledger-cosmos = "==1.57.0"
open-aea-ledger-ethereum = "==1.57.0"
open-aea-ledger-ethereum-flashbots = "==1.57.0"
open-aea-cli-ipfs = "==1.57.0"
clea = "==0.1.0rc4"
cytoolz = "==0.12.3"
docker = "6.1.2"
fastapi = "==0.110.0"
eth-hash = "==0.7.0"
eth-account = "==0.8.0"
eth-keyfile = "==0.6.1"
eth-keys = "==0.4.0"
eth-rlp = "==0.3.0"
eth-typing = "==3.5.2"
eth-utils = "==2.3.1"
eth-abi = "==5.1.0"
frozenlist = "==1.4.1"
hexbytes = "==0.3.1"
ipfshttpclient = "==0.8.0a2"
jsonschema = "==4.3.3"
multidict = "==6.0.5"
requests-toolbelt = "1.0.0"
starlette = "==0.36.3"
uvicorn = "==0.27.0"
web3 = "==6.1.0"
psutil = "^5.9.8"
pyinstaller = "^6.8.0"
aiohttp = "3.9.5"
halo = "^0.0.31"
icecream = "^2.1.3"

[tool.poetry.group.development.dependencies]
tomte = { version = "0.2.17", extras = ["cli"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
